Create tweetable scope
Define a scope on Tweet that returns all rows that have not been tweeted
yet and are approved, i.e. which are tweetable.

We don't want to send the same tweet again. And we also don't want to
tweet out bars that have not been approved yet as they might need
further review or editing.