Create a TweetService
This service determines when tweets are next due and then sends them to
Twitter. The next due calculation is based on the time between tweets.
Currently this is set to 24 hours. So 1 day after the last tweet was
sent a new one should go out. The `Tweet#tweeted_at` column is the
valule of reference here.

If no tweets have ever been sent `next_tweet_due` will also return true.

The `send_when_due` method can be called by Cron job regularly and only
actually send a tweet when it should.

Also this structure would allow for a potential change in the future to
vary the time between tweets based on the amount of tweets cued up to
send.