Change 'send' to 'publish'
This commit renames all place where we use the word send to the word
publish instead.

'Publish' more clearly conveys what happens with a tweet.

Interestingly this first became a problem when trying to use the shoulda
matcher gem which calls the Ruby method `#send`[1] on the Tweet model.
So not overwriting that standard method is another advantage to the rename.

[1] https://ruby-doc.org/core-2.6/Object.html#method-i-send