Capture the id of a tweet
When a tweet is sent the Twitter API returns the id of the tweet.
This commit makes sure it's stored in the databse.

The columnn needs to be a bigint to store the large number returned by
Twitter.

This should help us identify a tweet. With the id it can be found and
each record in the database pertaining to a bar should match a tweet
once it has been published.