Validate presence of Tweet#link_to_song and Tweet#link_to_lyrics
This also requires updating a host of tests that relied on Tweet objects
to be created as setup and which objects can now not be created if they
fail these new validations.

This update implicitly calls for the introduction for something like
FactoryBot which would have made the updating of these objects easier.