Follow up first tweet with a second tweet containing link to lyrics and song
Taking the response from the first tweet we can take the returned id to
send a response, i.e. a follow-up tweet. The follow-up should contain a
link to the song and the lyrics. Since validations check their presence
these should always exist for a Tweet record.

At this stage the id of the second tweet is not recorded. It should be
possible to find the responses to the first tweet.

However it's becoming more apparent that a refactor of the domain model
and code are in order soon. The relationship between bar and tweet could
be redefined and the tweeting and formatting of tweets does not
necessarily belong in the Tweet model. A `TweetPresenter` does already
exist and the `TweetService` could expand it's responsibility for
sending tweets.

At this stage it is ok as an MVP to test if this app gains any traction.