Add Cronjob to run rake task every five minutes and publish tweets when due
This cronjob runs every 5 minutes. It runs a rake task that in turn runs
rails code to publish tweets when they are due.

The cronjob follows the recommendations from Dokku:

http://dokku.viewdocs.io/dokku/deployment/one-off-processes/#general-cron-recommendations

It is important to remove the container after the job has run to not
leave container duplicates hanging around and clutter up disk space.